Fixed data types to avoid potential overflow issues when dealing with strings

+++ include/string.h
@@ -6,7 +6,7 @@
 /*   By: chlimous <chlimous@student.42.fr>	    +#+  +:+	   +#+	      */
 /*						  +#+#+#+#+#+	+#+	      */
 /*   Created: 2023/12/13 13:00:57 by chlimous	       #+#    #+#	      */
-/*   Updated: 2024/01/04 18:32:21 by chlimous         ###   ########.fr       */
+/*   Updated: 2024/01/07 21:42:41 by chlimous         ###   ########.fr       */
 /*									      */
 /* ************************************************************************** */
 
@@ -25,12 +25,12 @@ char	*ft_strchr(const char *s, int c);
 char	*ft_strrchr(const char *s, int c);
 int		ft_strncmp(const char *s1, const char *s2, size_t n);
 char	*ft_strnstr(const char *big, const char *little, size_t len);
-char	*ft_substr(char const *s, unsigned int start, size_t len);
+char	*ft_substr(char const *s, size_t start, size_t len);
 char	*ft_strjoin(char const *s1, char const *s2);
 char	**ft_split(char const *s, char const *sep_charset);
 char	*ft_strtrim(char const *s1, char const *set);
-char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
-void	ft_striteri(char *s, void (*f)(unsigned int, char *));
+char	*ft_strmapi(char const *s, char (*f)(size_t, char));
+void	ft_striteri(char *s, void (*f)(size_t, char *));
 char	*ft_strncpy(char *dest, char *src, size_t size);
 char	*ft_strcpy(char *dest, const char *src);
 char	*ft_strndup(const char *src, size_t size);
